backend/requirements.txt
# Core
Django>=4.2,<5.0
djangorestframework
django-cors-headers
psycopg2-binary
Pillow
python-decouple

# Async & Queue
celery
redis
channels
channels-redis==4.2.0
daphne
watchdog

# Auth
djangorestframework-simplejwt

# AI & ML
langchain
langchain-groq
scikit-learn==1.8.0
pefile
yara-python
virustotal3

# Hashing
blake3

# File Type Detection
python-magic-bin

# Network Analysis
psutil==5.9.8
geoip2
aiohttp

# Data
pandas==3.0.1
numpy

# PDF
xhtml2pdf
